changes to embedded compiler logic to fix compiler GUI debugger feature
The compiler GUI debugger feature was affected by changes to enable the
embedded compiler. Recent behavior in the compiler GUI debugger
include displaying internal queries when displaying a user query
and occasionally crashing the sqlci process. The changes to fix this
issue is to ensure that the GUI debugger is only enabled when the
(compiler) context of the user query is the active context. Another change is
to ensure that the GUI debugger is disabled and related structures are
appropriately reset when the user queries fails during compilation.

Change-Id: Ia75f889c2ed24efabc4501432cc6b5706094b3e6
Closes-Bug: #1339205