changes to embedded compiler logic to fix compiler GUI debugger feature
The compiler GUI debugger feature was affected by changes to enable the
embedded compiler. Recent behavior in the compiler GUI debugger
include displaying internal queries when displaying a user query
and occasionally crashing the sqlci process. The changes to fix this
issue is to ensure that the GUI debugger is only enabled when the
(compiler) context of the user query is the active context. Another change is
to ensure that the GUI debugger is disabled and related structures are
appropriately reset when the user queries fails during compilation.

fixes to the rentrant compiler logic to support spjs so external
compilers are avoided when compiling "call" statements. Changes
were made to file CmpSeabaseDDLtable.cpp method
CmpSeabaseDDL::getSeabaseRoutineDesc.

Closes-Bug: #1340960

Change-Id: Ia75f889c2ed24efabc4501432cc6b5706094b3e6
Closes-Bug: #1339205