Fix for bug 1317729: Disallow SALT clause on approximate numeric columns.
Trying to use the SALT clause on an approximate numeric column (float,
real double precision) will now return error 1120. The reason is that
rounding errors could cause us to compute the wrong SALT partition value,
and we might not find a row or allow duplicates. Note that it is not a
good idea to use an approximate numeric as a key column, although that
is still allowed.

Also fixed a bug with salted tables that use a STORE BY clause.

Change-Id: I70d76970fd4da5b2c73a0448c552c92bb446f478