Bug 1340646 SPJ with SQL query statements failed with SQLCODE 2005
Fixing typo in the commit message.

The problem is that the thread static global variable trafMDDescsInfo_
is deallocated after cmp context is deallocated, because trafMDDescsInfo_ is
allocated on CTXTHEAP.

Since trafMDDescsInfo_ is a thread global, changed to be allocated from
CLI Globals exCollHeap().  For all the data members also to be allocated
on exCollHeap(), the method processDDLandCreateDescs() now takes a new
parameter from where the output parameters are allocated from.

Change-Id: I441d23459389646f076ae706ef0559b8a5474c54