fix bug 1343661(cleanup HBASE partitioning),  bug 1347819 and bug 1343566
1343661: please refer to sql/sqlcomp/DefaultConstConstants.h for the
definition of CQD HBASE_PARTITIONING.
1347819: the change is in ExExeUtilLoad.cpp to disable HASH2 for fast load.
1343566: method NADefaults::getTotalNumOfESPsInCluster() now returns the
correct value if CQD PARALLEL_NUM_ESPS is set to an integer value.
Rework 1 to address Dave's review comments.
Rework 2 to address Khaled's review comments as follows. A Boolean flag
(isTrafLoadPrep_) is added to class BinWA to better control the type of
partitioning functions needed for the traf preparation step.  When we are
binding all nodes, the flag is set to TRUE which instructs createNAFileSet()
not to create hash2.

Change-Id: I9f886ce5361ff4ede16c592e8a3c4e5ca89d76b7