Avoid pushdown of sampling to HBase if oversampling is used
In general, when a Sample clause is present that specifies random
sampling, the sampling is pushed down to the HBase layer and performed
using a filter. When oversampling (sampling rate > 100%) is used, this
can not be done because the filter used is incapable of returning
multiple copies of a given row. The result would be to return all
selected rows once, instead of a greater number of rows that is
consistent with the oversampling.

This change opts out of sampling pushdown if the sample percentage is
greater than 100%, performing the sampling instead with a Sample
operator within Trafodion.

Note that the function RelSample::isSimpleRandomRelative() formerly
contained a comment stating that it returned FALSE if oversampling was
used, although this was not the case. Rather than adding that check to
the function, the comment was removed and the check for oversampling
was added to RelSample::bindNode() as a separate pushdown check.
isSimpleRandomRelative() is also used as a condition for parallel
execution, which should not be ruled out simply because random
oversampling is used.

Closes-Bug #1340432

Change-Id: I6894be2735cf3f7e5a5838bfcab4e1055a7274cd