Change to automatically set the HBase client side cache size.
Previously the size of HBase Scan cache size was set through a CQD.
With this change this cache size will be set to a value between [min, max]
based of the estimated number of rows accessed per scan node instance.
The values of min and max can be specified through cqds. The default values
are [100, 10000]. The previous default value was 100. The desired effect
of this change is to improve throughput for large scans. Simple experiments
are showing an improvement in the range of to 2-3X.

Patch Set 2:
Set cache size for update and delete too, this will be functional when
card estimate code is added later
Added stubs for estRowsAccess for Update and Delete
Changed name of cqds
Added cache size info to explain output
modularized code by putting logic to set cache size in a method

Change-Id: I5fac104666bd125671c818fc71ddad666f186ad7