Change to automatically set the HBase client side cache size.
Previously the size of HBase Scan cache size was set through a CQD.
With this change this cache size will be set to a value between [min, max]
based of the estimated number of rows accessed per scan node instance.
The values of min and max can be specified through cqds. The default values
are [100, 10000]. The previous default value was 100. The desired effect
of this change is to improve throughput for large scans. Simple experiments
are showing an improvement in the range of to 2-3X.

Change-Id: I5fac104666bd125671c818fc71ddad666f186ad7