Fix for loader issue with hive varchar into Trafodion datetime column
When loading a hive table that has a timestamp as a character into
a Trafodion target table that has a timestamp data type, we get an
error during optimization, complaining about the type incompatibility.
This is with the ALLOW_INCOMPATIBLE_ASSIGNMENT CQD set to ON.
One issue is that we don't set up the update to select value id map
correctly, this map needs to map the target timestamp to the
cast(source col as timestamp), not the character source column.

Change-Id: Id599c1f4d9a563003fd394f87c070f257d94c4a7