Undo part of yesterday's fix for bug 1343615
To get the main branch back to a usable state for performance testing,
we are undoing my change from yesterday:

4ce0ba57c97b720182f98d2a430a2f275212e1de
Bug 1343615: Duplicated rows for parallel scan on salted table

as well as changing the default value for CQD HBASE_HASH2_PARTITIONING
to OFF. To get parallel plans for HASH2 partitioned salted tables working,
we'll try UTT testing in the future.

Patch set 2:

Patch set 1 generated lots of cores, due to inconsistent checking in
the NATable cache for when we use HASH2 for salted tables. Changing
the CQDs to make them simpler and consistent with NATable cache:

  // HBASE_HASH2_PARTITIONING (ON - OFF for now):
  // - OFF: Treat salted table like a non-salted table
  // - ON: Always use HASH2 for salted table

  // HBASE_RANGE_PARTITIONING (ON):
  // - OFF: Never use HBase ranges for partitioning
  // - ON: Use HBase ranges (stats may override)

  // HBASE_STATS_PARTITIONING (ON):
  // - OFF: Don't use stats split
  // - ON: Use stats split if possible, except for HASH2

Note that HBASE_PARTITIONING goes away and is replaced with the
new HBASE_STATS_PARTITIONING CQD.

Also fixing test issues with compGeneral/TEST013 and executor/TEST013.
The optimizer switched from an index scan to an equally good table scan
which should not cause the test to fail.

Change-Id: Ie95bbdac6f8465d4f7fb74786b170d1bd895cefb