Fix bug #1328254 - Explain options 'm' got TRANSLATE error
This modifies ExplainTuple::setCol() so that when the code trims the
'explain' output to the maximum length for a line of explain output,
it is careful to *not* split any multi-byte UTF8 character that might
be at the end of line. It puts the '*' (indicator of continuation) over
the top of the *first* byte of the UTF8 character and a '\0' after the
asterisk.

This change has been pre-reviewed by Mike.

A unit test was developed which set EXPLAIN_DESCRIPTION_COLUMN_SIZE
so that a UTF8 character was put at the end of the output and thus
got the TRANSLATE error before putting in this fix.  The error did not
occur once this fix was put in.

Also, all dev regressions were run to ensure that the fix has no
side-effects.

Files changed:
sql/comexe/ExplainTuple.cpp

Change-Id: I977ce552951d9773fcad84dfcddac9657327842f