Support for SALTED index.
CREATE INDEX supports a new clause "SALT LIKE TABLE".
This causes SALT column to be leading column in the index
Duplicate columns in the index table are now eliminated. This is a bugfix.
Showddl and Invoke will show the SALT syntax and column respectively for
the index. CREATE index also supports HBASE_OPTIONS clause.
A bug seen when NULLABLE partitioning columns are used is also fixed.

Change-Id: If378ffca29ee83dd4b7928c784b8d34d76f50049