Fix for bug 1378089
The HTableClient java objects are removed from the cache upon ODBC/JDBC
disconnection. However, in use HTableClient objects are not garbage
collected because of the global reference being held on the JNI side.

This change is to flag the members of the HTableClient for garbage collection
to reduce the amount of leak in java objects. In-use HTableClient objects are
still being leaked. A complete fix will come later.

Change-Id: I10df40ad58cfde9f2c50e612df7c910cbf33726d