Fix for Bug 1378415 - Generator Internal error assertion failure
Prior to the fix, in HbaseAccess searchKey expressions, the _SALT_ column
is represented by its base table column. This may cause some problems
downstream during codeGen time (HbaseAccess:codeGen) where the vid
representing the _SALT_ column cannot be found in the maptable.
The fix changes the search key expression to use a veg on the _SALT_
column instead of the basecolumn.

Change-Id: I6a78cced65760444485a6e2fa10a700286f96ff9
Closes-Bug: #1378415