Use new HFIle location for HBase 0.98
The code to lookup the HFiles for a given table was changed
to accommodate their new location within the HDFS file system.
The location seems to have changed at some point between 0.94
and 0.98, although I could find no documentation of the change,
nor any HBase property that could be used to identify the
path to the filesystem node under which HFiles for tables
are stored. This change affects lookup of HFiles for the purpose
of estimating the number of rows in a table.

Change-Id: I17b892144c03a2e9cebd3de22838177ffb6a324d
Closes-Bug: 1379862