Fix of bug #1328730:
JDBC T4 getTables methods only returns leading 100 rows,
if the total results is larger than 100. The second fetch
of JDBC Driver gets nothing. MXOSRVR actually fetch all rows
in statments data buffer, but in the second fetch, server
sends the buffer starts at the same address as last fetch, in
which the leading bytes of previous fetch are all already set
to '0'.

To fix this, uses a offset in statment to mark the start address
for next fetch of the buffer.

This passes dev unit tests for this specific problem, and QA
regression tests.

Change-Id: Ifd098dec38acb0ca7ff0cec153328d3b1349f3e1