Update TOOLSDIR references for Hadoop, HBase, Hive dependencies
Update the section of Hadoop dependencies that references TOOLSDIR
locations. The purpose is to use the same build-time dependencies
regardless of the distribution installed on the build machine.

Build should work without having a distro installed and without using
install_local_hadoop before building. But if you have local_hadoop
installed, that context still gets preference over TOOLSDIR.

The installed-distro settings (HortonWorks, Cloudera, MapR) are used
for run-time dependencies.

TOOLSDIR are build-time references and are only used if we are in a
source tree. So if your machine has a distro installed and TOOLSDIR
defined, then you'll get build-time references if sqenvcom.sh finds a
Makefile, indicating we are in source tree.

Change-Id: I1b86bf8c454467c6adef15e66014fd6da59b1f15
Blueprint: infra-build-hadoop-deps