Set up security keys while constructing NATable
This change corrects a problem in NATable by setting up the
secKeySet_ ComSecurityKeySet while the NATable is constructed.
In this way, the NATable::resetAfterStatement method will not
flag it as a leak. The change also includes some temporary logic
to handle scenarios where authorization is not enabled, as well
as temporary code to detect NATables for the PRIVMGR_MD schema
and skip privilege checking for these. Finally, the change
includes a new regression test for REVOKE and query invalidation.

Change-Id: If24cae9998bd77acda94f6ec5a70b59eddc9ce1d
Closes-Bug: #1371199