Fix for bug in determining range partition boundary value
When the boundary value of a range partition function is determined by
decoding the encoded value stored as the split boundary in hbase, some mistakes
were being made. This problem was seen for nullable varchar columns. The
symptom is crash during compile with an overflowing stack. The problem
was that the buffer provided to the runtime code to decode the encoded value
was laid out incorrectly. The null bytes were interchanges with the var length
indicator bytes. Thanks to Hans suggesting t his fix.

Change-Id: Id0bd2b187500d283860d0d12ecb9d8d743b429e9