Fix for bug 1392090
Earlier JNI objects in Trafodion were made thread independent. HBaseClient_JNI
object, the creator/user of other Trafodion JNI objects were not able to used
in a different thread because the thread specific JNIEnv was not setup. This
change initializes the JNIENV * when it is not initialized in the thread.

Change-Id: I827002a495c3373b66fb2c88f80675d9da188193