fix cardinality errors for better DoP for OE
Problems fixed are as follows.
1.	TableAnalysis::getLocalPredsOnPrefixOfList() prematurely breaks
out of the loop to collect predicates on key columns.
2.	TableDesc::getSaltColumnAsSet() does not return the result in
the form of base columns valueIds.
3.	When the dop is estimated lower than the low bound of ESPs, the
low bound value is used.
4.	SimpleFileScanOptimizer::computeSingleSubsetSize() incorrectly
applies the computed predicates on salted column which reduces the
subset size by CQD HIST_NO_STATS_UEC (default to 2). A subset size
specifies the # of rows processed.

Change-Id: I85bebf0d31d8f3e2db5bfe7c00cb4467b53308ac