Computed column key predicates for MDAM
Moving the generation of computed column predicates out of the
SearchKey logic and making it available as a static method on
class ScanKey. This allows us to compute these predicates before
we create the Disjuncts data structure that is used in a file
scan, where it will go into a SearchKey or an MdamKey.

Also fixing a bug that stopped after the first predicate found
on a computed column, so it failed to produce both a begin and
and end key value when selecting a range of values
(removed a "break" in ScanKey::createComputedColumnPredicates)

Change-Id: I4297d789ded8522eb67d5441ac281657ff90e774