Fix for several security/QI bugs
Fix for bug 1396774 & bug 1396746:

When indexes are dropped as part of drop table, cache entries are not
being removed correctly from the compiler process(es)

Fix for BY <owner> create table bug:

When using the undocumented BY <owner> clause for CREATE TABLE and an
unregistered user name is specified, the object owner is incorrectly
set to 0.  Once 0, a subsequent drop table fails without giving an error

MANAGE_LIBRARY support:

Require a user to be assigned the MANAGE_LIBRARY component privilege in
order to create a library.  This is in addition to other privileges.

Change-Id: I0379b1ab9cd26c334c5a4343577aec27a8d43401