Miscellaneous minor changes
1) Revert the previous change to ComTdbHbaseAccess::displayRowId() for
SHOWPLAN. The displayed row IDs are prepared in
HbaseAccess::genListsOfRows. For non-missing IDs, they should be in
NAString format with length and ascii data. Any problem found in
displayRowId should be traced back before the generator code.

2) Remove the part of comments in sql/sqlcomp/DefineConstants.h that
requires adding new CQD names to the end of the constant list. This is
no longer mandatory in Trafodion.

3) Enable NA memory overflow checking when running executor and qat
suites in debug flavor. This would increase the run time for these two
suites by about 10%.

Change-Id: I854ec776fa5fb815c39677d469f8f46f18d51402