Manageability changes - event mgmt and stats publication
Implements changes to support event management using log4cpp.
Configuration files are located in $MY_SQROOT/conf folder and all logs
files are located in $MY_SQROOT/logs folder
For more information see the blueprint at:
https://blueprints.launchpad.net/trafodion/+spec/eventmanagement

Implements changes for publication of statistics to repository. For more
information see the blueprint at:
https://blueprints.launchpad.net/trafodion/+spec/repositorymetrics

Note:
In this initial delivery publication of statistics is disabled by
default and it can be enabled via DCS property. This code has been
reviewed internally prior to merging with mainline

Documentation:
https://wiki.trafodion.org/wiki/index.php/Trafodion_Manageability

Included timestamp to be part of the primarykey for metric aggregation
table

Addressed some of the comments and incorporated Anoop's change for
repository

Change-Id: I517575233c10b2a8683cdd1d53a2eec96d7c2a6f