Fixed typos in user-visible messages and sample code
Fixed spellings and minor grammar in strings used in
in output, and sample code supplied to users.

One possibly harmful misspelling was found in source code,
file sqf/monitor/linux/montim.cxx used "Contianer_ExitProcess"
where it should have used "Container_ExitProcess".

Five files which could be changed for this reason are being held back
because they are also in open change 875.  They will be checked in after
that change is merged to the main branch:

  conn/odbc/src/odbc/dependencies/linux/smxoevl.h
  conn/trafci/src/org/trafodion/ci/InterfaceQuery.java
  sqf/src/sp_proxy/sp_common.cpp
  sqf/src/tm/tmlibmsg.h
  sqf/src/tm/tmtimer.cpp

Change-Id: I3761e9e1518ab39415806bba0f4ed93d14e1d41c