Fix for an issue with drop schema statement
Drop schema was not removing the NATable cached entry for the tables
in the schema because catalog name and object name got switched in an
internal statement.

When the SSCP process dies while it had locked the RMS semaphore,
SSMP didn't take control of the semaphore to release it. There was no activity
in SSMP process and hence SSMP didn't take control. Code changed to register
SSCP process in the shared segment and this can result in death message to
be processed by SSMP process. Processing of death message of SSCP will allow
SSMP to take control and release the semaphore held by SSCP when it died.

Hence enabled seabase/TEST025 regression

Change-Id: Ib4bc93b9614d5222033487091e3a2461f72e11f0