Fix for hang encountered during hive tests.
Hive tests have been hanging for quite a while on the official slave
machines.
The cause for this particular hang was that the main thread was
destroying the cursor (ExLobCursor) object . The worker thread was kind
of slow and it continues to access the cursor after the main thread had
destroyed it. pthread calls exhibit undefined behavior when
uninitialized mutex or condition variables are accessed. Added code to
prevent this timing issue in the worker threads.

Added trace utlitlty to diagnose hangs and execution issues.
The trace messages look like this and get logged into trace file on the
local directory named trace_threads.<pid>. It is controlled by an
environment variable TRACE_HDFS_THREAD_ACTIONS.The envvar is checked
only ones and the file handle is checked each time the trace needs to be
done. If file handle is NULL, no message is logged.

Change-Id: I95519b61d71339c719e37500bccae111ce070a15