Fix LP bug 1323865 -ODBC api test fails with Unknown PCode instruction
During PCODE generation, we were attempting to generate a PCODE
instruction to compare two operands for equality.  The two operands
had a data type of REC_BYTE_V_ASCII_LONG which is used only by
ODBC.  PCIT::getMemoryAddressingMode() does not currently know how
to handle that datatype so it returned AM_NONE for the operand
type.  That resulted in a failure later.

Fix was to detect the operand(s) of that datatype and call
ex_clause::pCodeGenerate(...) rather than doing PCODE generation
of the current expression.

Note: Also found a line saying
      return ex_clause::pCodeGenerate(space, f);
      which has been missing for a long time.  We got away with it
      because the preceding 'if' was always false for Trafodion.

Files changed:
.../exp/ExpPCodeClauseGen.cpp

Change-Id: I441f5b57fb1b5a67f4e4872d308efe0af5d320db