Fix core during dcsstop when repository is enabled
When the repository feature is enabled at time MXOSRVR may core during a
dcsstop/sqstop. This is because the repository based thread does get
terminated properly before the sqstop is invoked. Fix is to call the
SyncPublicationThread() during terminateThreads() where the rest of the
threads in MXOSRVR are being stopped.

Change-Id: Ib7b5de187ff02da49974b206832e25218891119c