Adding more run-time memory allocations from NAHeap
This set of changes moves some of the string vector variables in HBase
access operators from standard string template to our NAList and
NAString (or HbaseStr for row IDs). In the process, allocationis of the
objects will be from our HAHeap instead of the system heap. This would
help us tracking memory usage and detecting leaks easier.

In addition, a change in ExHbaseAccessTcb::setupListOfColNames()
prevents unnecessary allocations to populate the columns list unless it
is empty. The Google profiling tools helped us on identifying this
problem.

also, removed ExHbaseAccessDeleteTcb operator which was not used.

Change-Id: I87ab674ab8e3d291f2fc9563718d88de537ae96b