Misleading error when Hivestats can't be created
When Update Statistics is performed on a Hive table, the
trafodion.hivestats schema is created if it doesn't already
exist. If the user issuing the ustat statement does not
have the create_schema component privilege, this will fail, but
the code as it existed ignored the error, assuming it was
a "schema already exists" error, which is expected in most
cases. If the schema does not exist, an error is then given
when an attempt is made to create the histograms table in
the non-existing schema.

This has been changed to check the error when the schema
creation attempt fails, so the correct error is generated.
This is a partial fix, because a user with manage_statistics
privilege should not be prevented from executing Update
Statistics just because the trafodion.hivestats schema did
not already exist. The likely full solution would be to
create the schema automatically at Trafodion initialization.

Change-Id: I26561100bab9f6a59f3f5d0d56ce9d2aab380fa3
Partial-Bug: #1425748